Move order methods to digital module, add category creation on seed fix translation.
